backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
pillow==10.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1 